Add CSEM prior models
This commit adds prior models to Example0, Example1 and Example2. These prior
models are almost the same as the true models, just without the resistive body.

The goal of the CSEM branch was to combine both MT and CSEM and to be able to
use both data types/methods to resolve structures at different depths, however,
at the moment only CESM or MT can be used.